% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{analyzeEdgeListCat}
\alias{analyzeEdgeListCat}
\title{Performs significance association of a subnetwork with a field.}
\usage{
analyzeEdgeListCat(edge_indexes, osa, net, field, correction = "bonferroni",
  alternative = "greater")
}
\arguments{
\item{edge_indexes}{The index of the edges in the network that comprise the module.}

\item{osa}{The ordered sample annotation data frame.}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided",
"greater" or "less". You can specify just the initial letter.}

\item{correction.}{The method to apply for multiple testing correction. Valid values are identical
to those available to the p.adjust function.  The default is to
apply 'bonferroni' correction.}
}
\description{
Performs significance association of a subnetwork with a field.
}
