% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{analyzeNetCat}
\alias{analyzeNetCat}
\title{Performs significance association for every edge in the network.}
\usage{
analyzeNetCat(net, osa, field, correction = "bonferroni")
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{osa}{The ordered sample annotation data frame.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{correction.}{The method to apply for multiple testing correction. Valid values are identical
to those available to the p.adjust function.  The default is to
apply 'bonferroni' correction.}
}
\description{
Performs significance association for every edge in the network.
}
