% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{clusterEdges}
\alias{clusterEdges}
\title{Perfoms heirarchical clustering of edges in a network based on their sample compositions.}
\usage{
clusterEdges(net, distMethod = "manhattan", hclustMethod = "ward.D2")
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{distMethod}{The method to be provided to the dist function.}

\item{hclustMethod}{The method to be provided to the hclust method.}
}
\value{
An object of class hclust which describes the tree produced by the
   clustering process.
}
\description{
This function uses the dist function to calucate a distance and the
hclust function to generate the dendrogram.
}
