% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{drawEdgeListHeatMap}
\alias{drawEdgeListHeatMap}
\title{Draws a heatmap with dendrogram of a module in the netework.}
\usage{
drawEdgeListHeatMap(edge_indexes, osa, net, fieldOrder)
}
\arguments{
\item{edge_indexes}{The index of the edges in the network that comprise the module.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{fieldOrder}{A vector containing a list of sample attribute names for reordering of
of the samples. Each element of this vector should be the name of
a column in the osa matrix.  The sorting occurs first by the first
element, then by the second, etc.}
}
\description{
Draws a heatmap with dendrogram of a module in the netework.
}
