% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{drawEdgeTreeHeatMap}
\alias{drawEdgeTreeHeatMap}
\title{Draws a heatmap with dendrogram with clusters identified.}
\usage{
drawEdgeTreeHeatMap(sampleMatrix, tree, osa, fieldOrder, image_name = NA)
}
\arguments{
\item{sampleMatrix}{The sample matrix as created by the getSampleMatrix() function.}

\item{tree}{An instance of an hclust object as created by the clusterEdges() function.}

\item{osa}{The sample annotation matrix as created by the loadSampleAnnotations()
function.}

\item{fieldOrder}{A vector containing a list of sample names for reordering of
of the samples. Each element of this vector should be the name of
a column in the osa matrix.  The sorting occurs first by the first
element, then by the second, etc.}

\item{image_name}{The filename for saving the image. If no name is provided then
the image is not saved.}
}
\description{
Draws a heatmap with dendrogram with clusters identified.
}
