% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{fishers_test}
\alias{fishers_test}
\title{Performs a fisher's exact test on a sample string.}
\usage{
fishers_test(category, sample_types, sample_ref, correction = NA,
  alternative = "greater")
}
\arguments{
\item{category}{The annotation category.}

\item{sample_types}{The ordered array of annotations for all samples (i.e. a full column
of the sample annotation matrix).}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided",
"greater" or "less". You can specify just the initial letter.}

\item{rep_samples}{An array of numeric values indicating the strength of the sample in
the edge(s).}

\item{correction.}{The method to apply for multiple testing correction. Valid values are identical
to those available to the p.adjust function.  The default is to
apply no correction.}
}
\description{
The sample annotation matrix consists of multiple columns of annotation
types (or fields).  Each annotation type may have multiple distinct
values. For example the type may be 'sex' and the categories may be
'male' and 'female'.  This function performs a Fisher's test on
a single category amongst all the other categories within the annotation
type.
}
