% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT.R
\name{FoldChangeEdgeCat}
\alias{FoldChangeEdgeCat}
\title{Fold Change Edge}
\usage{
FoldChangeEdgeCat(i, net, ematrix)
}
\arguments{
\item{i}{The index of the edge in the network}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.
This function is intended to be run on a network that has already been
reduced by analyzeNetCat.}

\item{ematrix}{The expression matrix data frame.}
}
\description{
Determines the fold change between the in and out group of each edge. 
Returns 0 if there is not enough power to do test
}
\details{
Determines fold change through determining the abs(mean) of the in and out group, and dividing
the larger abs(mean) by the smaller one.
}
