% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT.R
\name{FoldChangeNetCat}
\alias{FoldChangeNetCat}
\title{FoldChange Net}
\usage{
FoldChangeNetCat(net, ematrix, progressBar = TRUE)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.
This function is intended to be run on a network that has already been
reduced by analyzeNetCat.}

\item{ematrix}{The expression matrix data frame.}

\item{progressBar}{Set to FALSE to repress progress bar}
}
\description{
Determines fold change through determining the abs(mean) of the in and out group, and dividing
the larger abs(mean) by the smaller one.
}
\details{
Does this for your network of interest
}
