% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{getEdgeExpressionMatrix}
\alias{getEdgeExpressionMatrix}
\title{Creates an Edge Expression Matrix (EEM) from a network data frame.}
\usage{
getEdgeExpressionMatrix(net, osa, ematrix)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{osa}{The ordered sample annotation data frame.}

\item{ematrix}{The expression matrix data frame.}
}
\value{
An nxm*2 dataframe contianing the EEM
}
\description{
The EEM is an n x m*2 array were n is the number of samples in the annotation matrix and
m is the number of edges in the network. Each edge is represented in the matrix by two columns
with each containing gene expression of the source and target genes respectively.
}
