% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{getPairGMMEdges}
\alias{getPairGMMEdges}
\title{Performs pair-wise Gaussian Mixture Models (GMM) on a pair of genes}
\usage{
getPairGMMEdges(source, target, ematrix, th = 0.5, method = "spearman")
}
\arguments{
\item{source}{The name of the source gene.}

\item{target}{The name of the target gene.}

\item{ematrix}{The expression matrix data frame.}

\item{th}{The similarity value threshold. Any correlation value greater than or equal to the
absolute value of this number is kept as an edge.}

\item{method}{The correlation method to perform (see the cor function for valid options)}
}
\value{
Returns a data frame of edges in the same format as the network returned by
  the loadNetwork function.
}
\description{
Performs pair-wise Gaussian Mixture Models (GMM) on a pair of genes
}
