% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{getSampleStringArray}
\alias{getSampleStringArray}
\title{Converts and edge sample string into an array.}
\usage{
getSampleStringArray(i, net)
}
\arguments{
\item{i}{The index of the edge in the network}

\item{net}{A network dataframe containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}
}
\value{
An array whose values are  1 if the sample is in the cluster,
  0 if the sample is in another cluster, 9 if the sample is missing and
  7 if the sample was removed as a pairwise outlier or 8 if the sample
  was removed as a cluster outlier.
}
\description{
Converts and edge sample string into an array.
}
