% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{graphEdgeList}
\alias{graphEdgeList}
\title{Draws the graph of a set of edges in the network}
\usage{
graphEdgeList(edge_indexes, net, sim_col = "sc", osa = data.frame(),
  field = NA)
}
\arguments{
\item{edge_indexes}{The index of the edges in the network that comprise the module.}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}
}
\description{
Draws the graph of a set of edges in the network
}
