% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{graphNet}
\alias{graphNet}
\title{Draws the graph of the entire network}
\usage{
graphNet(net, osa = data.frame(), sim_col = "sc")
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}
}
\description{
Draws the graph of the entire network
}
