% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{highlightEdgeList}
\alias{highlightEdgeList}
\title{Colors the edges of a given edge list on the currently plotted graph.}
\usage{
highlightEdgeList(edge_indexes, graph, layout, net, color)
}
\arguments{
\item{edge_indexes}{The index of the edges in the network that comprise the module.}

\item{graph}{The igraph object as returned from graphNet or graphEdgeList.}

\item{layout}{The layout object as returned from graphNet or graphEdgeList}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{color}{The color that should be used to highlith the edges}
}
\description{
Colors the edges of a given edge list on the currently plotted graph.
}
