% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{loadKINCNetwork}
\alias{loadKINCNetwork}
\title{Imports a network file produced by KINC into a data frame.}
\usage{
loadKINCNetwork(network_file, KINC_version = "1.0")
}
\arguments{
\item{network_file}{The full path to the network file on the file system.}
}
\value{
A dataframe containing the network file contents
}
\description{
Imports a network file produced by KINC into a data frame.
}
