% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{loadSampleAnnotations}
\alias{loadSampleAnnotations}
\title{Imports sample annotations.}
\usage{
loadSampleAnnotations(annotation_file, sample_header = "Sample")
}
\arguments{
\item{annotation_file}{The path to the file containing the annotations.}

\item{smaple_order}{A file containing the order of samples in the sample strings of
the network.  The file should contain the list of samples each on
a separate line.}
}
\value{
A data frame containing the annotations in the order of the samples.
}
\description{
The format of the sample annotation file is tab delimited where the
first column is the sample name and every other column contains
the annotation.  The file must contain a header with the annotation
types and the first column of the header should be the word 'Sample'
}
