% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mergeClusters}
\alias{mergeClusters}
\title{The recursive merging function called by mergeCommunities().}
\usage{
mergeClusters(cedges, cnodes, th = 0.5)
}
\description{
The recursive merging function called by mergeCommunities().
}
