% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mergeCommunities}
\alias{mergeCommunities}
\title{Merges linked community clusters that have the most similar sets of nodes.}
\usage{
mergeCommunities(lc)
}
\arguments{
\item{lc}{A linkcomm object.}
}
\value{
A list were each element of the list is the
  set of nodes and edges of the merged clusters.
}
\description{
This merging function recursively iterates through all of the clusters
and performs a pair-wse Jaccard comparision between all clusters.
Those with the highest Jaccard score that are above the given
threshold (i.e. deafult of 0.5) are candidates for merging.
}
\details{
This is a helper function for the findLinkedCommunities() function and is
not meant to be called on its own.
}
