% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{performClusterBinomialTest}
\alias{performClusterBinomialTest}
\title{Performs a fisher's exact test on a set of samples.}
\usage{
performClusterBinomialTest(category, field, i, net, osa, ematrix,
  t1_psucc = 0.25, t2_psucc = 0.75, verbose = FALSE)
}
\arguments{
\item{category}{The annotation category to be used for testing. It must be a valid
category in the field if the osa identified by the field argument.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{i}{The index of the edge in the network}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{ematrix}{The expression matrix data frame.}

\item{t1_psucc}{The probabilty of success for the first binomial test.  This value
indicates the percentage of samples in the cluster that must
belong to the category.}

\item{t2_psucc}{The probability of success for the second binomial test. This value
indicates the percentage of category samples that must belong
to the cluster.}
}
\description{
An annotation field and a specific category are provided. This
test reports if the category is enriched within the cluster.
It does not indiciate if the category is singificantly more
prominent in the cluster. For that, try the sampleClusterBTest
function.
}
