% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT.R
\name{performClusterTTest}
\alias{performClusterTTest}
\title{Performs a weltch's t-test}
\usage{
performClusterTTest(i, net, ematrix, num_not_category, num_category,
  cluster_osa_indexes, non_cluster_osa_indexes)
}
\arguments{
\item{i}{The index of the edge in the network}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{ematrix}{The expression matrix data frame.}

\item{num_not_category}{The number of samples that are not in the cluster}

\item{num_category}{The number of samples that are in the cluster}

\item{cluster_osa_indexes}{indexes of the samples that are in the cluster}

\item{non_cluster_osa_indexes}{indexes of the samples that are not in the cluster

Power Analysis
Determine if we have the power to distinguish if there is a difference with our number of samples
If we do not have the power for one or both of the genes, we can not determine if it is signicantly different
return 1 if this is true}
}
\description{
Performs a weltch's t-test
}
