% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot2DEdgeList}
\alias{plot2DEdgeList}
\title{Plots a 2D scatterplot for a given list of edges.}
\usage{
plot2DEdgeList(edge_indexes, osa, net, ematrix, field = NA,
  samples = NA)
}
\arguments{
\item{edge_indexes}{The index of the edge in the network data frame.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{ematrix}{The expression matrix.}

\item{field}{The field in the sample annotation matrix by which to split the points
into separate layers.}

\item{samples}{An array of sample names for which the scatterplot should be included in the plot.
If no value is provided then all samples with values are included.}
}
\description{
Plots a 2D scatterplot for a given list of edges.
}
