% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot3DEdgeList}
\alias{plot3DEdgeList}
\title{Plots a 3D scatterplot for a given list of edges.}
\usage{
plot3DEdgeList(edge_indexes, osa, net, ematrix, field, colors = NA,
  samples = NA)
}
\arguments{
\item{edge_indexes}{The index of the edge in the network data frame.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{ematrix}{The expression matrix.}

\item{field}{The field in the sample annotation matrix by which to split the points
into separate layers.}

\item{colors}{An array of colors per sample.}

\item{samples}{Limit the plot to only the samples indexes provided.}
}
\description{
Plots a 3D scatterplot for a given list of edges.
}
