% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot3DPair}
\alias{plot3DPair}
\title{Plots a 3D scatterplot for a given pair of nodes.}
\usage{
plot3DPair(i, j, osa, ematrix, field, colors = NA, samples = NA)
}
\arguments{
\item{i}{The index of the first node in the expression matrix.}

\item{j}{The index of the second node in the expression matrix.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{ematrix}{The expression matrix.}

\item{field}{The field in the sample annotation matrix by which to split the points
into separate layers.}

\item{colors}{An array of colors per sample.}
}
\description{
Plots a 3D scatterplot for a given pair of nodes.
}
