% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plotEdgeList}
\alias{plotEdgeList}
\title{Plots a 2D scatterplot for a given list of edges.}
\usage{
plotEdgeList(edge_indexes, net, ematrix, cor_col = "sc", colors = NA,
  samples = NA)
}
\arguments{
\item{edge_indexes}{The index of the edge in the network data frame.}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{ematrix}{The expression matrix.}

\item{cor_col}{Then name of the column in the net argument that provides the similarity score.}

\item{colors}{An array of colors per sample.}

\item{samples}{An array of sample names for which the scatterplot should be included in the plot.
If no value is provided then all samples with values are included.}
}
\description{
Plots a 2D scatterplot for a given list of edges.
}
