% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plotGene}
\alias{plotGene}
\title{Plots a scatterplot of gene expression across a set of samples.}
\usage{
plotGene(gene, ematrix, osa, field, xlab = field)
}
\arguments{
\item{gene}{The name of the gene to plot}

\item{ematrix}{The expression matrix.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{field}{The name of the field by which to order the points along the
x-axis.}

\item{xlab}{The label for the x-axis of the plot.}
}
\description{
Samples will be ordered by the values of field.
}
