% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{sampleClusterFTest}
\alias{sampleClusterFTest}
\title{Performs a fisher's exact test on a set of samples.}
\usage{
sampleClusterFTest(category, field, osa, net, ematrix, cluster_samples,
  alternative = "greater")
}
\arguments{
\item{category}{The annotation category to be used for testing. It must be a valid
category in the field if the osa identified by the field argument.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{ematrix}{The expression matrix data frame.}

\item{cluster_samples}{The indexe of samples wihin the cluster. These correspond to indexes
in the column names of the expression matrix.}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided",
"greater" or "less". You can specify just the initial letter.}

\item{correction.}{The method to apply for multiple testing correction. Valid values are identical
to those available to the p.adjust function.  The default is to
apply 'hochberg' correction.}
}
\description{
An annotation field and a speicfic category are provided. This
test reports if the category is enriched within the cluster.
It does not indiciate if the category is singificantly more
promenint in the cluster. For that, try the sampleCluterBTest
function.
}
