% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{sampleClusterlogit}
\alias{sampleClusterlogit}
\title{Performs logistic regression for a set of Cluster samples
This uses the}
\usage{
sampleClusterlogit(category, field, osa, net, ematrix, cluster_samples)
}
\description{
Performs logistic regression for a set of Cluster samples
This uses the
}
