% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{saveKINCNetwork}
\alias{saveKINCNetwork}
\title{Exports a network data frame as a KINC compatible file.}
\usage{
saveKINCNetwork(net, network_file)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{network_file}{The path to which the network file will be saved.}
}
\description{
Exports a network data frame as a KINC compatible file.
}
