% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT.R
\name{TtestEdgeCat}
\alias{TtestEdgeCat}
\title{T-test Edge}
\usage{
TtestEdgeCat(i, net, ematrix)
}
\arguments{
\item{i}{The index of the edge in the network}

\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.
This function is intended to be run on a network that has already been
reduced by analyzeNetCat.}

\item{ematrix}{The expression matrix data frame.}
}
\description{
Performs a weltch's t-test on the two genes of an edge. Returns 1 if there is not enough power to do test,
otherwise will return t-test stat for both genes
}
