% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT.R
\name{TtestNetCat}
\alias{TtestNetCat}
\title{Performs a weltch's t-test on a network}
\usage{
TtestNetCat(net, ematrix, progressBar = TRUE)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.
This function is intended to be run on a network that has already been
reduced by analyzeNetCat.}

\item{ematrix}{The expression matrix data frame.}

\item{progressBar}{Set to FALSE to repress progress bar}
}
\description{
Performs a weltch's t-test on a network
}
