% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{analyzeEdgeCat}
\alias{analyzeEdgeCat}
\title{Performs significant testing of a single edge in the network for a set of annotation categories.}
\usage{
analyzeEdgeCat(i, osa, net, ematrix, field, test = "binomial",
  category = NA, t1_psucc = 0.25, t2_psucc = 0.75)
}
\arguments{
\item{i}{The index of the edge in the network}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{net}{A network dataframe containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{ematrix}{The expression matrix data frame.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{test}{The statistical test to perform: 'fishers' or 'binomial'}

\item{category}{By default this function will cylce through all categories of the
'field' argument. But to limit the analysis to just one category set it here.}

\item{t1_psucc}{For the bionomial test only.
The probabilty of success for the first binomial test.  This value
indicates the percentage of samples in the cluster that must
belong to the category.}

\item{t2_psucc}{For the binomial test only.
The probability of success for the second binomial test. This value
indicates the percentage of category samples that must belong
to the cluster.}
}
\description{
Performs significant testing of a single edge in the network for a set of annotation categories.
}
\examples{

}
