% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{analyzeEdgeCat}
\alias{analyzeEdgeCat}
\title{Performs significant testing of a single edge in the network for a set of annotation categories.}
\usage{
analyzeEdgeCat(i, osa, net, ematrix, field, test = "fishers",
  samples = c())
}
\arguments{
\item{i}{The index of the edge in the network}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{net}{A network dataframe containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{ematrix}{The expression matrix data frame.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{test}{The statistical test to perform.  This is 'fishers' for an enrichment
test.}

\item{samples}{Limit the analysis to only the samples indexes provided.}
}
\description{
Performs significant testing of a single edge in the network for a set of annotation categories.
}
\examples{

}
