% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{analyzeNetCat}
\alias{analyzeNetCat}
\title{Performs significant testing of each edge in the network for a set of annotation categories.}
\usage{
analyzeNetCat(
  net,
  osa,
  ematrix,
  field,
  test = "binomial",
  correction = "hochberg",
  progressBar = TRUE,
  category = NA,
  t1_psucc = 0.25,
  t2_psucc = 0.75
)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{osa}{The ordered sample annotation data frame.}

\item{ematrix}{The expression matrix data frame.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{test}{The statistical test to perform.  This is 'fishers' for an enrichment
test.}

\item{progressBar}{Set to FALSE to repress progress bar}

\item{category}{By default this function will cylce through all categories of the
'field' argument. But to limit the analysis to just one category set it here.}

\item{t1_psucc}{For the bionomial test only.
The probabilty of success for the first binomial test.  This value
indicates the percentage of samples in the cluster that must
belong to the category.}

\item{t2_psucc}{For the binomial test only.
The probability of success for the second binomial test. This value
indicates the percentage of category samples that must belong
to the cluster.}

\item{correction.}{The method to apply for multiple testing correction. Valid values are identical
to those available to the p.adjust function.  The default is to
apply 'hochberg' correction.}

\item{nthreads}{The number of threads (CPU cores/threads) used to analyze the network.}
}
\description{
Performs significant testing of each edge in the network for a set of annotation categories.
}
