% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{analyzeNetCat}
\alias{analyzeNetCat}
\title{Performs significant testing of each edge in the network for a set of annotation categories.}
\usage{
analyzeNetCat(net, osa, ematrix, field, test = "fishers",
  correction = "hochberg", samples = c(), progressBar = TRUE)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{osa}{The ordered sample annotation data frame.}

\item{ematrix}{The expression matrix data frame.}

\item{field}{The field in the osa variable on which enrichment will be performed.}

\item{test}{The statistical test to perform.  This is 'fishers' for an enrichment
test.}

\item{samples}{Limit the analysis to only the samples indexes provided.}

\item{progressBar}{Set to FALSE to repress progress bar}

\item{correction.}{The method to apply for multiple testing correction. Valid values are identical
to those available to the p.adjust function.  The default is to
apply 'hochberg' correction.}
}
\description{
Performs significant testing of each edge in the network for a set of annotation categories.
}
