% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{findLinkedCommunities}
\alias{findLinkedCommunities}
\title{Finds Linked communities in the network.}
\usage{
findLinkedCommunities(net, file_prefix, module_prefix = "M",
  sim_col = "sc", hcmethod = "complete", meta = TRUE,
  ignore_inverse = TRUE)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}

\item{file_prefix}{A prefix to add to the beginning of each file name.}

\item{module_prefix}{A prefix to add to the beginning of the module names. By deafult this is simply
the letter 'M'.}

\item{sim_col}{The column name in the network data frame for the similarity score. Default is 'sc',}

\item{hcmethod}{A character string naming the hierarchical clustering method to use. Can be one of
"ward.D", "single", "complete", "average", "mcquitty", "median", or "centroid".
Defaults to "single".}

\item{meta}{Indicates if modules should be collapsed into meta-modules. If set to TRUE then
the linked communities returned are meta modules. Defaults to FALSE.}
}
\value{
The linked communities object.
}
\description{
This function generates three output files that it writes to the current
working directory.
}
