% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{getEdgeSamples}
\alias{getEdgeSamples}
\title{Uses the edge sample string to return the list of sample indexes in the edge.}
\usage{
getEdgeSamples(i, net)
}
\arguments{
\item{i}{The index of the edge in the network}

\item{net}{A network dataframe containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}
}
\value{
An array of sample indexes that are part of the cluster. Because these
  indexes are derived from the edge sample string they indexes should
  correspond to the order of samples in the expression matrix.
}
\description{
Uses the edge sample string to return the list of sample indexes in the edge.
}
