% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMM.R
\name{getPairGMMEdges}
\alias{getPairGMMEdges}
\title{Performs pair-wise Gaussian Mixture Models (GMM) on a pair of genes}
\usage{
getPairGMMEdges(
  source,
  target,
  ematrix,
  minc = 30,
  method = "spearman",
  th = 0.5,
  plot = FALSE,
  iterations = 5
)
}
\arguments{
\item{source}{The name of the source gene.}

\item{target}{The name of the target gene.}

\item{ematrix}{The expression matrix data frame.}

\item{minc}{The minimum number of samples that must be present to identify a cluster. Default = 30.}

\item{method}{The correlation method to perform (see the cor function for valid options)}

\item{th}{The similarity value threshold. Any correlation value greater than or equal to the
absolute value of this number is kept as an edge. Derfault = 0.5.}

\item{plot}{Boolean indicating if the Rmixmod clusters scatterplot should be drawn. Default = FALSE}

\item{iterations}{The number of times to execute the GMM models.  The model with the fewest clusters from
all the iterations will be used.}
}
\value{
Returns a data frame of edges in the same format as the network returned by
  the loadNetwork function.
}
\description{
Performs pair-wise Gaussian Mixture Models (GMM) on a pair of genes
}
