% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{getSampleMatrix}
\alias{getSampleMatrix}
\title{Generates a matrix containing edge sample strings digits.}
\usage{
getSampleMatrix(net)
}
\arguments{
\item{net}{A network data frame containing the KINC-produced network.  The loadNetwork
function imports a dataframe in the correct format for this function.}
}
\value{
A matrix
}
\description{
The returned matrix contains as many rows as there are edges
in the network and as many columns as their are samples. Each
column corresponds to a sample in the same order as the
sample strings from the network.
}
