% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mergeClusters}
\alias{mergeClusters}
\title{The recursive merging function called by mergeCommunities().
This is a helper function for the findLinkedCommunities() and mergeCommunities
functions and is not meant to be called on its own.}
\usage{
mergeClusters(cedges, cnodes, th)
}
\description{
The recursive merging function called by mergeCommunities().
This is a helper function for the findLinkedCommunities() and mergeCommunities
functions and is not meant to be called on its own.
}
