% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{plot3DPair}
\alias{plot3DPair}
\title{Plots a 3D scatterplot for a given pair of nodes.}
\usage{
plot3DPair(
  gene1,
  gene2,
  osa,
  ematrix,
  field,
  colors = NA,
  samples = NA,
  highlight = c()
)
}
\arguments{
\item{gene1}{The name of the first gene in the pair.}

\item{gene2}{The name of the second gene in the pair.}

\item{osa}{The sample annotation matrix. One column must contain the header 'Sample'
and the remaining colums correspond to an annotation type.  The rows
of the anntation columns should contain the annotations.}

\item{ematrix}{The expression matrix.}

\item{field}{The field in the sample annotation matrix by which to split the points
into separate layers.}

\item{colors}{An array of colors per sample.}
}
\description{
Plots a 3D scatterplot for a given pair of nodes.
}
